ZNHY_developer/requirements.txt
flask==2.3.3
python-docx==0.8.11
requests==2.31.0
openai==0.28.0
python-dotenv==1.0.0